Backend/User_service_v1/User_service/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
canvasapi==3.3.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.26.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.30.6
google-auth==2.36.0
python-jose==3.3.0
httpx==0.28.0
google-cloud-pubsub==2.27.1
boto3 == 1.35.72
python-decouple== 3.8
jwt==1.3.1
PyJWT==2.10.1